requirements.txt
pandas
streamlit
psycopg2
numpy
plotly.express
sqlalchemy 
streamlit_antd_components